+++ include/GafferScene/Duplicate.h
@@ -103,21 +103,15 @@ class GAFFERSCENE_API Duplicate : public BranchCreator
 
 	private :
 
-		// The BranchCreator::parentPlug() must be used to define the place where the duplicates
-		// are to be parented, but it's much more natural for the user to simply specify which object
-		// they want to duplicate, and expect that the duplicates will appear alongside the original.
-		// This output plug is used to compute the appropriate parent from the target, and is connected
-		// into BranchCreator::parentPlug() so that the user doesn't need to worry about it.
-		Gaffer::StringPlug *outParentPlug();
-		const Gaffer::StringPlug *outParentPlug() const;
+		IE_CORE_FORWARDDECLARE( DuplicatesData );
 
-		// We need the list of names of the duplicates in both computeBranchChildNames()
-		// and computeBranchTransform(), so we compute it on this intermediate plug so that
-		// the list is cached and the work is shared between the two methods.
-		Gaffer::InternedStringVectorDataPlug *childNamesPlug();
-		const Gaffer::InternedStringVectorDataPlug *childNamesPlug() const;
+		// Used to store the names and transforms for each copy. Must be
+		// evaluated in a context where `scene:path` is one of the source
+		// locations.
+		Gaffer::ObjectPlug *duplicatesPlug();
+		const Gaffer::ObjectPlug *duplicatesPlug() const;
 
-		void branchSource( const ScenePath &branchPath, ScenePath &source ) const;
+		void branchSource( const ScenePath &sourcePath, const ScenePath &branchPath, ScenePath &source ) const;
 
 		static size_t g_firstPlugIndex;